WOAH-FUNCTION FIN
So they aren't all perfect. But I've created functions that correspond to requested equations and so forth. I have my doubts about the EAP function but since it returns values and I only have a couple hours left, I think it's best to move forward with setting up the skeleton for a package!

AND AWAAAAAAY WE GO